Factor out ItemType.ts
