+++ src/pages/non-modal-ingredient-page/non-modal-ingredient-page.tsx
@@ -2,21 +2,7 @@ import React from 'react';
 import { useParams } from 'react-router-dom';
 import { useSelector } from 'react-redux';
 import styles from './non-modal-ingredient-page.module.css';
-
-type TItemType = {
-  _id: string;
-  name: string;
-  type: string;
-  proteins: number;
-  fat: number;
-  carbohydrates: number;
-  calories: number;
-  price: number;
-  image: string;
-  image_mobile: string;
-  image_large: string;
-  __v: number;
-}
+import TItemType from '../../types/ItemType';
 
 function NonModalIngredientPage() {
   const { id } = useParams();
