+++ src/types/ItemType.ts
@@ -1,0 +1,17 @@
+type TItemType = {
+    _id: string;
+    name: string;
+    type: string;
+    proteins: number;
+    fat: number;
+    carbohydrates: number;
+    calories: number;
+    price: number;
+    image: string;
+    image_mobile: string;
+    image_large: string;
+    __v: number;
+    uniqueId?: string;
+};
+
+export default TItemType;
